refactor(pipeline): remove lib/discover.ts (replaced by phases/)
- splitWaysByConnectivity moved to lib/way-connectivity.ts
- namesAreSimilar moved to lib/name-similarity.ts
- Existing ski-filter test now imports isSkiOnlyWay from lib/ski-filter.ts
  directly (the re-export intermediary is gone with lib/discover.ts)
